Fix numpy-pytorch conversion error for miss forest imputation.

+++ deep_generative_models/imputation/miss_forest_imputation_task.py
@@ -34,7 +34,7 @@ class MissForestImputation(BasicImputation):
         imputed = model.transform(scaled_inputs)
 
         # go back to torch (annoying)
-        return torch.from_numpy(imputed)
+        return torch.from_numpy(imputed).float()
 
 
 if __name__ == '__main__':
